+++ wasm/src/programs/snapshot_query.rs
@@ -112,3 +112,49 @@ impl QueryTrait<CurrentNetwork> for SnapshotQuery {
         Ok(self.block_height)
     }
 }
+
+
+mod snapshot_helpers {
+    use super::*;
+    type StateRootNative = <CurrentNetwork as snarkvm_console::network::Network>::StateRoot;
+
+    pub fn collect_commitments_from_trace<T>(
+        _trace: &T,
+    ) -> Result<Vec<Field<CurrentNetwork>>, String> {
+        Ok(vec![])
+    }
+
+    pub async fn build_snapshot_query(
+        node_url: &str,
+        commitments: &[Field<CurrentNetwork>],
+    ) -> Result<SnapshotQuery, String> {
+        let (state_root, block_height) = snapshot_head(node_url).await?;
+        let mut query = SnapshotQuery::new(block_height, &state_root_to_string(state_root))?;
+
+        for c in commitments {
+            let c_str = c.to_string();
+            let sp_str = fetch_state_path_at_root(node_url, &c_str, &state_root_to_string(state_root)).await?;
+            query.add_state_path(&c_str, &sp_str)?;
+        }
+        Ok(query)
+    }
+
+    // Prefer an API that returns both in one call; otherwise fetch root then height immediately.
+    async fn snapshot_head(node_url: &str) -> Result<(StateRootNative, u32), String> {
+        // TODO: call existing client / REST: (root, height)
+        Err("snapshot_head() not implemented".into())
+    }
+
+    async fn fetch_state_path_at_root(
+        node_url: &str,
+        commitment: &str,
+        state_root: &str,
+    ) -> Result<String, String> {
+        // TODO: call endpoint that returns a StatePath string for (commitment, state_root)
+        Err("fetch_state_path_at_root() not implemented".into())
+    }
+
+    fn state_root_to_string(root: StateRootNative) -> String {
+        root.to_string()
+    }
+}
